Remove Update Indexes button
