+++ src/app/containers/App/App.js
@@ -1,13 +1,25 @@
 import React from 'react';
-import NavBar from '../NavBar/NavBar';
+import { unwrapResult } from '@reduxjs/toolkit';
+import { useDispatch, useSelector } from 'react-redux';
+import { refreshIndexes, updateIndexes } from '../../reducers/indexSlice';
 import Stocks from '../Stocks/Stocks';
 import Indexes from '../Indexes/Indexes';
 import Profile from '../Profile/Profile';
 
 function App() {
+  const dispatch = useDispatch();
+  const indexes = useSelector((state) => state.indexes);
+
+  if (indexes['^GSPC'].price === '') {
+    dispatch(refreshIndexes(Object.keys(indexes).join(',')))
+      .then(unwrapResult)
+      .then((originalPromiseResult) => {
+        dispatch(updateIndexes(originalPromiseResult));
+      });
+  }
+
   return (
     <div className="App">
-      <NavBar />
       <Indexes />
       <Stocks />
       <Profile />
